requirements.txt
streamlit
numpy
pandas
plotly-express
openpyxl